docker/requirements.txt
Flask==3.1.0
Flask-Migrate==4.0.7
alembic==1.14.0
SQLAlchemy==2.0.36
config==0.5.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
